Client submodule works, untested

+++ productswindow.cpp
@@ -1,9 +1,5 @@
 #include "productswindow.h"
 #include "ui_productswindow.h"
-#include "sql.h"
-#include <QtSql>
-#include <QtDebug>
-#include <QObject>
 
 productsWindow::productsWindow(QWidget *parent) :
     QDialog(parent),
@@ -91,18 +87,21 @@ void productsWindow::on_button_removeSelected_clicked()
     if(ui->combo_select->currentText()=="ID"){
         qDebug() << "Nie wybrano rekordu!";
     }else{
-        int id = ui->combo_select->currentText().toInt();
-        //qDebug() << id;
-        sql conn;
-        conn.dbOpen();
-        QSqlQuery *query = new QSqlQuery(conn.db);
-
-        query->prepare("delete from produkty where id=:id");
-        query->bindValue(":id",id);
-        query->exec();
-        conn.dbClose();
-
-        updateProductList();
+        QMessageBox::StandardButton reply;
+        reply = QMessageBox::question(this, "Usuwanie rekordu", "Czy napewno chcesz usunąć ten rekord?", QMessageBox::Yes|QMessageBox::No);
+        if (reply == QMessageBox::Yes) {
+            int id = ui->combo_select->currentText().toInt();
+            sql conn;
+            conn.dbOpen();
+            QSqlQuery *query = new QSqlQuery(conn.db);
+            query->prepare("delete from produkty where id=:id");
+            query->bindValue(":id",id);
+            query->exec();
+            conn.dbClose();
+            updateProductList();
+        } else {
+            updateProductList();
+        }
     }
 }
